Restructure some code and data structures.

+++ nest/route.h
@@ -222,7 +222,7 @@ typedef struct rte {
 #endif
 #ifdef CONFIG_BABEL
     struct {
-      byte metric;			/* Babel metric */
+      u16 metric;			/* Babel metric */
       struct babel_entry *entry;
     } babel;
 #endif
@@ -380,7 +380,7 @@ typedef struct rta {
 #define RTS_OSPF_EXT2 10		/* OSPF external route type 2 */
 #define RTS_BGP 11			/* BGP route */
 #define RTS_PIPE 12			/* Inter-table wormhole */
-#define RTS_BABEL 13			/* Inter-table wormhole */
+#define RTS_BABEL 13			/* Babel route */
 
 #define RTC_UNICAST 0
 #define RTC_BROADCAST 1
